+++ src/apis/index.ts
@@ -16,9 +16,11 @@
  */
 
 import { asArray, cloneObj, normalizeString, toBooleanSafe, toStringSafe } from '../index';
+import { getCpuUsage, getDiskSpace } from '../system';
 import * as _ from 'lodash';
 import { Response } from 'express';
 import { readFile, readFileSync } from 'fs-extra';
+import * as os from 'os';
 import * as Path from 'path';
 
 /**
@@ -34,15 +36,6 @@ export interface ApiError {
      */
     en: string;
 }
-
-/**
- * A possible value for 'importApiErrors()' and 'importApiErrorsSync()' functions.
- *
- * If STRING: The path to the JSON file to import.
- * If BUFFER: The binary content as UTF-8 JSON data.
- * If OBJECT or ARRAY: One or more items to import.
- */
-export type ImportApiErrorsArgument = string | Buffer | ApiErrorWithKey | ApiErrorWithKey[];
 
 /**
  * An entry for an API error with a key.
@@ -76,6 +69,55 @@ export interface ApiResult {
 }
 
 /**
+ * Options for 'createMonitoringApiResult()' function.
+ */
+export interface CreateMonitoringApiResultOptions {
+    /**
+     * An optional function, which checks a database connection.
+     */
+    databaseConnectionChecker?: () => boolean | PromiseLike<boolean>;
+}
+
+/**
+ * A possible value for 'importApiErrors()' and 'importApiErrorsSync()' functions.
+ *
+ * If STRING: The path to the JSON file to import.
+ * If BUFFER: The binary content as UTF-8 JSON data.
+ * If OBJECT or ARRAY: One or more items to import.
+ */
+export type ImportApiErrorsArgument = string | Buffer | ApiErrorWithKey | ApiErrorWithKey[];
+
+/**
+ * A result of a 'createMonitoringApiResult()' function call.
+ */
+export interface MonitoringApiResult {
+    /**
+     * The CPU usage in percentage.
+     */
+    cpu_load: number;
+    /**
+     * Indicates if a database connection is established or available.
+     */
+    database_connected?: boolean;
+    /**
+     * The total disk space, in bytes.
+     */
+    disk_space: number;
+    /**
+     * The total disk space in use, in bytes.
+     */
+    disk_space_used: number;
+    /**
+     * The total ram, in bytes.
+     */
+    ram: number;
+    /**
+     * The ram in use, in bytes.
+     */
+    ram_used: number;
+}
+
+/**
  * Additional options for 'sendResponse()' function.
  */
 export interface SendResponseOptions {
@@ -105,6 +147,94 @@ function applyApiErrors(errors: ApiErrorWithKey[]) {
 
         API_ERRORS[KEY] = cloneObj(CLONED_ERR);
     }
+}
+
+/**
+ * Creates an object for an result of a monitoring API endpoint.
+ *
+ * @param {CreateMonitoringApiResultOptions} [opts] Custom options.
+ *
+ * @return {Promise<MonitoringApiResult>} The promise with the result (object).
+ */
+export async function createMonitoringApiResult(
+    opts?: CreateMonitoringApiResultOptions
+): Promise<MonitoringApiResult> {
+    if (_.isNil(opts)) {
+        opts = <any>{};
+    }
+
+    let cpu_load: number;
+    try {
+        cpu_load = await getCpuUsage();
+    } catch { } finally {
+        if (isNaN(cpu_load)) {
+            cpu_load = -1;
+        }
+    }
+
+    let database_connected: boolean;
+    try {
+        const CONNECTION_CHECKER = opts.databaseConnectionChecker;
+        if (!_.isNil(CONNECTION_CHECKER)) {
+            database_connected = toBooleanSafe(
+                await Promise.resolve(
+                    CONNECTION_CHECKER()
+                )
+            );
+        }
+    } catch {
+        database_connected = false;
+    }
+
+    let disk_space: number;
+    let disk_space_used: number;
+    try {
+        const DS = await getDiskSpace();
+
+        disk_space = DS.total;
+        disk_space_used = DS.used;
+    } catch { } finally {
+        if (isNaN(disk_space)) {
+            disk_space = -1;
+        }
+
+        if (isNaN(disk_space_used)) {
+            disk_space_used = -1;
+        }
+    }
+
+    let ram: number;
+    let ram_used: number;
+    try {
+        ram = parseInt(
+            toStringSafe(
+                os.totalmem()
+            ).trim()
+        );
+
+        ram_used = ram - parseInt(
+            toStringSafe(
+                os.freemem()
+            ).trim()
+        );
+    } catch { } finally {
+        if (isNaN(ram)) {
+            ram = -1;
+        }
+
+        if (isNaN(ram_used)) {
+            ram_used = -1;
+        }
+    }
+
+    return {
+        cpu_load: cpu_load,
+        database_connected: database_connected,
+        disk_space: disk_space,
+        disk_space_used: disk_space_used,
+        ram: ram,
+        ram_used: ram_used,
+    };
 }
 
 /**